AI_Assignment/requirements.txt
transformers
torch
requests
beautifulsoup4
pandas
scikit-learn
python-dotenv
streamlit
